fix file loading bugs

+++ main.h
@@ -188,7 +188,7 @@ void bitwise_or(CPUState *state, uint8_t operand_rd, uint8_t operand_rn);
 uint8_t memory_access(CPUState *state, uint8_t reg, uint16_t address, uint8_t mode, uint8_t srcDest);
 bool hasChanged(int* lastValue, int currentValue);
 
-void setupMmap(CPUState *state, uint8_t program_size);
+void setupMmap(CPUState *state, size_t program_size);
 bool handleWrite(CPUState *state, uint16_t address, uint8_t value);
 
 void mmuControl(CPUState *state, uint8_t value);
